% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geerate_explanations.R
\docType{data}
\name{available_methods}
\alias{available_methods}
\title{Acronyms for available explanation methods.}
\format{
An object of class \code{character} of length 5.
}
\usage{
available_methods
}
\description{
Acronyms for available explanation methods.
}
\keyword{datasets}
