% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_smoothed_gradients}
\alias{calculate_smoothed_gradients}
\title{Calculates smooth gradients for a CNN.}
\usage{
calculate_smoothed_gradients(
  model,
  input_imgs,
  preprocessing_function,
  class_index,
  num_samples,
  noise_sd
)
}
\arguments{
\item{model}{Tensorflow model.}

\item{input_imgs}{Input images if for of 4-D tensor.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}

\item{class_index}{Class index. If set to `NULL` index with max predicted probability will be selected.}

\item{num_samples}{Number of noised samples per one image.}

\item{noise_sd}{Gaussian noise standard deviation.}
}
\value{
smooth gradients for a CNN.
}
\description{
Calculates smooth gradients for a CNN.
}
