% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_class_indexes}
\alias{check_class_indexes}
\title{Checks if `class_index` argument is correct.}
\usage{
check_class_indexes(input_imgs, class_index)
}
\arguments{
\item{input_imgs}{Input images in form of 4-D tensor.}

\item{class_index}{Class index.}
}
\description{
Selects predictions form output based on indexes.
}
