% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainer.R
\name{CNNexplainer}
\alias{CNNexplainer}
\title{Creates `CNNexplainer` object.}
\value{
`CNNexplainer` object.
}
\description{
Creates `CNNexplainer` object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CNNexplainer$new()}}
\item \href{#method-show_available_methods}{\code{CNNexplainer$show_available_methods()}}
\item \href{#method-explain}{\code{CNNexplainer$explain()}}
\item \href{#method-clone}{\code{CNNexplainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplainer$new(iput_data, model, preprocessing_function)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_available_methods"></a>}}
\if{latex}{\out{\hypertarget{method-show_available_methods}{}}}
\subsection{Method \code{show_available_methods()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplainer$show_available_methods()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-explain"></a>}}
\if{latex}{\out{\hypertarget{method-explain}{}}}
\subsection{Method \code{explain()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplainer$explain(
  class_index,
  methods,
  num_samples,
  noise_sd,
  steps,
  patch_size,
  absolute_values,
  grayscale
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
