% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_explanations.R
\name{CNNexplanations}
\alias{CNNexplanations}
\title{Creates `CNNexplanations` object.}
\value{
`CNNexplanations` object.
}
\description{
Creates `CNNexplanations` object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CNNexplanations$new()}}
\item \href{#method-print}{\code{CNNexplanations$print()}}
\item \href{#method-get_explanations}{\code{CNNexplanations$get_explanations()}}
\item \href{#method-get_metadata}{\code{CNNexplanations$get_metadata()}}
\item \href{#method-combine}{\code{CNNexplanations$combine()}}
\item \href{#method-plot_and_save}{\code{CNNexplanations$plot_and_save()}}
\item \href{#method-clone}{\code{CNNexplanations$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes `CNNexplanations` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplanations$new(explanations, ids, target_sizes, methods, input_imgs_paths)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{explanations}}{Explanations.}

\item{\code{ids}}{Explainers ids.}

\item{\code{target_sizes}}{Target sizes list.}

\item{\code{methods}}{Explanation methods.}

\item{\code{input_imgs_paths}}{Input images paths.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints basic information about object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplanations$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_explanations"></a>}}
\if{latex}{\out{\hypertarget{method-get_explanations}{}}}
\subsection{Method \code{get_explanations()}}{
Collects explanations from `CNNexplanations` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplanations$get_explanations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Collects metadata from `CNNexplanations` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplanations$get_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-combine"></a>}}
\if{latex}{\out{\hypertarget{method-combine}{}}}
\subsection{Method \code{combine()}}{
Combines `CNNexplanations` objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplanations$combine(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{`CNNexplanations` objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_and_save"></a>}}
\if{latex}{\out{\hypertarget{method-plot_and_save}{}}}
\subsection{Method \code{plot_and_save()}}{
Generates raster image(s) with explanations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplanations$plot_and_save(combine_plots, output_path = NULL, plot = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{combine_plots}}{Should images be combined.}

\item{\code{output_path}}{Where to save explanation plots.}

\item{\code{plot}}{Should explanation be plotted.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplanations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
