% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_plot_functions.R
\name{combine_cnn_multimodel_explanation_plots}
\alias{combine_cnn_multimodel_explanation_plots}
\title{Combines multimodel explanation plots.}
\usage{
combine_cnn_multimodel_explanation_plots(explanation_plots, n_models, n_imgs)
}
\arguments{
\item{explanation_plots}{Explanation plots.}

\item{n_models}{Number of models.}

\item{n_imgs}{Number of images.}
}
\value{
Raster image(s) with explanations.
}
\description{
Combines multimodel explanation plots.
}
