% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_plot_functions.R
\name{create_cnn_explanation_plots}
\alias{create_cnn_explanation_plots}
\title{Generates raster image(s) with explanations.}
\usage{
create_cnn_explanation_plots(explanations)
}
\arguments{
\item{explanations}{Explanations.}
}
\value{
Raster image(s) with explanations.
}
\description{
Generates raster image(s) with explanations.
}
