% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_methods.R
\name{filter_methods_by_network}
\alias{filter_methods_by_network}
\title{Filters available methods by type.}
\usage{
filter_methods_by_network(network_type)
}
\arguments{
\item{network_type}{Network type.}
}
\value{
Filtered `tibble` with available method for a network type.
}
\description{
Filters available methods by type.
}
