% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_last_conv2d_layer}
\alias{find_last_conv2d_layer}
\title{Finds last Conv2D layer in the network.}
\usage{
find_last_conv2d_layer(model)
}
\arguments{
\item{model}{Tensorflow model.}
}
\value{
Last Conv2D layer in the network.
}
\description{
Finds last Conv2D layer in the network.
}
