% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_methods.R
\name{find_method_name}
\alias{find_method_name}
\title{Finds explanation method name.}
\usage{
find_method_name(method)
}
\arguments{
\item{method}{Method acronym.}
}
\value{
Method full name.
}
\description{
Finds explanation method name.
}
