% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnn_explanations.R
\name{generate_cnn_explanations}
\alias{generate_cnn_explanations}
\title{Generates explanations for images.}
\usage{
generate_cnn_explanations(
  model,
  input_imgs_paths,
  id,
  preprocessing_function = NULL,
  class_index = NULL,
  methods = c("V", "GI", "SG", "SGI", "IG", "GB", "OCC", "GGC"),
  batch_size = length(input_imgs_paths),
  num_samples = 5,
  noise_sd = 0.1,
  steps = 20,
  patch_size = c(50, 50),
  grayscale = TRUE
)
}
\arguments{
\item{model}{Tensorflow model.}

\item{input_imgs_paths}{Input images paths.}

\item{id}{Explainer id.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}

\item{class_index}{Class index. If set to `NULL` index with max predicted probability will be selected.}

\item{methods}{Methods to be calculated.}

\item{batch_size}{Batch size.}

\item{num_samples}{Number of noised samples per one image.}

\item{noise_sd}{Gaussian noise standard deviation.}

\item{steps}{Integration steps. Must be positive integer.}

\item{patch_size}{Patch size. 2-D `integer` vector.}

\item{grayscale}{Boolean. Should gradients be converted from RGB to grayscale.}
}
\value{
Explanations for images.
}
\description{
Generates explanations for images.
}
