% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geerate_explanations.R
\name{generate_explanations}
\alias{generate_explanations}
\title{Generates explanations for images.}
\usage{
generate_explanations(
  model,
  input_imgs_paths,
  preprocessing_function = NULL,
  class_index = NULL,
  methods = available_methods,
  num_samples = 5,
  noise_sd = 0.1,
  steps = 20,
  absolute_values = TRUE,
  grayscale = TRUE
)
}
\arguments{
\item{model}{Tensorflow model.}

\item{input_imgs_paths}{Input images paths.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}

\item{class_index}{Class index. If set to `NULL` index with max predicted probability will be selected.}

\item{methods}{Methods to be calculated.}

\item{num_samples}{Number of noised samples per one image.}

\item{noise_sd}{Gaussian noise standard deviation.}

\item{steps}{Integration steps. Must be positive integer.}

\item{absolute_values}{Boolean. If `TRUE` absolute values of gradients will be returned.}

\item{grayscale}{Boolean. Should gradients be converted from RGB to grayscale.}
}
\value{
Explanations for images.
}
\description{
Generates explanations for images.
}
