% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_noisy_images}
\alias{generate_noisy_images}
\title{Generates noisy images.}
\usage{
generate_noisy_images(input_imgs, num_samples, noise_sd)
}
\arguments{
\item{input_imgs}{Input images if for of 4-D tensor.}

\item{num_samples}{Number of noised samples per one image.}

\item{noise_sd}{Gaussian noise standard deviation.}
}
\value{
Noised images.
}
\description{
Generates noisy images.
}
