% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_predictions_based_on_indexes}
\alias{get_model_predictions_based_on_indexes}
\title{Selects predictions form output based on indexes.}
\usage{
get_model_predictions_based_on_indexes(preds, class_index)
}
\arguments{
\item{preds}{Prediction tensor.}

\item{class_index}{Class index.}
}
\value{
Predictions form output based on indexes
}
\description{
Selects predictions form output based on indexes.
}
