% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guided_backpropagation.R
\name{guided_backpropagation}
\alias{guided_backpropagation}
\title{Calculates guided backpropagation for a CNN.}
\usage{
guided_backpropagation(
  model,
  input_imgs,
  preprocessing_function = NULL,
  class_index = NULL,
  absolute_values = TRUE,
  grayscale = TRUE,
  standardize = TRUE
)
}
\arguments{
\item{model}{Tensorflow model.}

\item{input_imgs}{Input images if for of 4-D tensor.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}

\item{class_index}{Class index. If set to `NULL` index with max predicted probability will be selected.}

\item{absolute_values}{Boolean. If `TRUE` absolute values of gradients will be returned.}

\item{grayscale}{Boolean. Should gradients be converted from RGB to grayscale.}

\item{standardize}{Boolean. Should gradients be standardized.}
}
\value{
Guided backpropagation for a CNN.
}
\description{
Calculates guided backpropagation for a CNN.
}
