% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guided_backpropagation.R
\name{guidedRelu}
\alias{guidedRelu}
\title{Guided ReLU with custom gradient function.}
\usage{
guidedRelu(x)
}
\arguments{
\item{x}{Layer input.}
}
\value{
Guided ReLU with custom gradient function.
}
\description{
Guided ReLU with custom gradient function.
}
