% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_gradients.R
\name{integrated_gradients}
\alias{integrated_gradients}
\title{Calculates integrated gradients for a CNN.}
\usage{
integrated_gradients(
  model,
  input_imgs,
  preprocessing_function = NULL,
  class_index = NULL,
  steps = 20,
  absolute_values = TRUE,
  grayscale = TRUE,
  standardize = TRUE
)
}
\arguments{
\item{model}{Tensorflow model.}

\item{input_imgs}{Input images if for of 4-D tensor.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}

\item{class_index}{Class index. If set to `NULL` index with max predicted probability will be selected.}

\item{steps}{Integration steps. Must be positive integer.}

\item{absolute_values}{Boolean. If `TRUE` absolute values of gradients will be returned.}

\item{grayscale}{Boolean. Should gradients be converted from RGB to grayscale.}

\item{standardize}{Boolean. Should gradients be standardized.}
}
\value{
integrated gradients for a CNN.
}
\description{
Calculates integrated gradients for a CNN.
}
