% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnn_explanations.R
\name{occlusion}
\alias{occlusion}
\title{Calculates occlusion sensitivity for a CNN.}
\usage{
occlusion(
  model,
  input_imgs,
  preprocessing_function = NULL,
  class_index = NULL,
  patch_size = c(50, 50)
)
}
\arguments{
\item{model}{Tensorflow model.}

\item{input_imgs}{Input images if for of 4-D tensor.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}

\item{class_index}{Class index. If set to `NULL` index with max predicted probability will be selected.}

\item{patch_size}{Patch size. 2-D `integer` vector.}
}
\value{
Occlusion sensitivity for a CNN.
}
\description{
Calculates occlusion sensitivity for a CNN.
}
