% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{per_image_standardization}
\alias{per_image_standardization}
\title{Per image standardization.}
\usage{
per_image_standardization(images)
}
\arguments{
\item{images}{Images.}
}
\value{
Standardized images.
}
\description{
Per image standardization.
}
