% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_explanations}
\alias{plot_explanations}
\title{Generates raster image.}
\usage{
plot_explanations(explanations, combine_plots = TRUE)
}
\arguments{
\item{explanations}{Explanations.}

\item{combine_plots}{Should images be combined.}
}
\value{
Raster image(s).
}
\description{
Generates raster image.
}
