% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_plot_functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Generates heatmap image.}
\usage{
plot_heatmap(main_data, main_grayscale, plot_data)
}
\arguments{
\item{main_data}{`data.frame` with `x`, `y` coordinates and color values.}

\item{main_grayscale}{Should images be plotted in grayscale.}

\item{plot_data}{`data.frame` with `x`, `y` coordinates and color values.}
}
\value{
Heatmap image.
}
\description{
Generates heatmap image.
}
