% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{save_cnn_explanation_plots}
\alias{save_cnn_explanation_plots}
\title{Plots raster image(s) with explanations.}
\usage{
save_cnn_explanation_plots(explanation_plots, combine_plots)
}
\arguments{
\item{explanation_plots}{Explanation plots.}

\item{combine_plots}{Should images be combined.}
}
\description{
Generates raster image(s) with explanations.
}
