% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnn_explanations.R
\name{smooth_grad}
\alias{smooth_grad}
\title{Calculates smooth-grad for a CNN.}
\usage{
smooth_grad(
  model,
  input_imgs,
  preprocessing_function = NULL,
  class_index = NULL,
  num_samples = 5,
  noise_sd = 0.1
)
}
\arguments{
\item{model}{Tensorflow model.}

\item{input_imgs}{Input images if for of 4-D tensor.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}

\item{class_index}{Class index. If set to `NULL` index with max predicted probability will be selected.}

\item{num_samples}{Number of noised samples per one image.}

\item{noise_sd}{Gaussian noise standard deviation.}
}
\value{
smooth-grad for a CNN.
}
\description{
Calculates smooth-grad for a CNN.
}
