% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_and_standarize_images}
\alias{transform_and_standarize_images}
\title{Transforms and standardizes gradients.}
\usage{
transform_and_standarize_images(
  gradients,
  absolute_values,
  grayscale,
  standardize
)
}
\arguments{
\item{gradients}{Tensor with gradients.}

\item{absolute_values}{Boolean. If `TRUE` absolute values of gradients will be returned.}

\item{grayscale}{Boolean. Should gradients be converted from RGB to grayscale.}

\item{standardize}{Boolean. Should gradients be standardized.}
}
\value{
Transformed and standardized gradients.
}
\description{
Transforms and standardizes gradients.
}
