% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_rgb_images_to_grayscale}
\alias{transform_rgb_images_to_grayscale}
\title{Transforms RGB images to grayscale.}
\usage{
transform_rgb_images_to_grayscale(rgb_images)
}
\arguments{
\item{rgb_images}{RGB images.}
}
\value{
Grayscale images.
}
\description{
Transforms RGB images to grayscale.
}
