% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_explainer.R
\name{CNNexplainer}
\alias{CNNexplainer}
\title{Creates `CNNexplainer` object.}
\value{
`CNNexplainer` object.
}
\description{
Creates `CNNexplainer` object.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{Tensorflow model.}

\item{\code{preprocessing_function}}{Image preprocessing function.}

\item{\code{id}}{Explainer id.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CNNexplainer$new()}}
\item \href{#method-show_available_methods}{\code{CNNexplainer$show_available_methods()}}
\item \href{#method-explain}{\code{CNNexplainer$explain()}}
\item \href{#method-clone}{\code{CNNexplainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes `CNNexplainer` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplainer$new(model, preprocessing_function, id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Tensorflow model.}

\item{\code{preprocessing_function}}{Image preprocessing function.}

\item{\code{id}}{Explainer id.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_available_methods"></a>}}
\if{latex}{\out{\hypertarget{method-show_available_methods}{}}}
\subsection{Method \code{show_available_methods()}}{
Prints available explanation methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplainer$show_available_methods()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-explain"></a>}}
\if{latex}{\out{\hypertarget{method-explain}{}}}
\subsection{Method \code{explain()}}{
Generates explanations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplainer$explain(
  input_imgs_paths,
  class_index,
  methods,
  num_samples,
  noise_sd,
  steps,
  patch_size,
  absolute_values,
  grayscale
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_imgs_paths}}{Input images paths.}

\item{\code{class_index}}{Class index. If set to `NULL` index with max predicted probability will be selected.}

\item{\code{methods}}{Methods to be calculated.}

\item{\code{num_samples}}{Number of noised samples per one image.}

\item{\code{noise_sd}}{Gaussian noise standard deviation.}

\item{\code{steps}}{Integration steps. Must be positive integer.}

\item{\code{patch_size}}{Patch size. 2-D `integer` vector.}

\item{\code{absolute_values}}{Boolean. If `TRUE` absolute values of gradients will be returned.}

\item{\code{grayscale}}{Boolean. Should gradients be converted from RGB to grayscale.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Explanations for images.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CNNexplainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
