% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnn_explanations.R
\name{gradient_x_input}
\alias{gradient_x_input}
\title{Calculates gradientXinput for a CNN.}
\usage{
gradient_x_input(vanilla_gradients, input_imgs, preprocessing_function)
}
\arguments{
\item{vanilla_gradients}{Vanilla gradients.}

\item{input_imgs}{Input images if for of 4-D tensor.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}
}
\value{
gradientXinput for a CNN.
}
\description{
Calculates gradientXinput for a CNN.
}
