% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnn_explanations.R
\name{guided_backpropagation}
\alias{guided_backpropagation}
\title{Calculates guided backpropagation for a CNN.}
\usage{
guided_backpropagation(
  model,
  input_imgs,
  preprocessing_function = NULL,
  class_index = NULL
)
}
\arguments{
\item{model}{Tensorflow model.}

\item{input_imgs}{Input images if for of 4-D tensor.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}

\item{class_index}{Class index. If set to `NULL` index with max predicted probability will be selected.}
}
\value{
Guided backpropagation for a CNN.
}
\description{
Calculates guided backpropagation for a CNN.
}
