% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnn_explanations.R
\name{integrated_gradients}
\alias{integrated_gradients}
\title{Calculates integrated gradients for a CNN.}
\usage{
integrated_gradients(
  model,
  input_imgs,
  preprocessing_function = NULL,
  class_index = NULL,
  steps = 20
)
}
\arguments{
\item{model}{Tensorflow model.}

\item{input_imgs}{Input images if for of 4-D tensor.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}

\item{class_index}{Class index. If set to `NULL` index with max predicted probability will be selected.}

\item{steps}{Integration steps. Must be positive integer.}
}
\value{
integrated gradients for a CNN.
}
\description{
Calculates integrated gradients for a CNN.
}
