% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Generates heatmap image.}
\usage{
plot_heatmap(plot_data)
}
\arguments{
\item{plot_data}{`data.frame` with `x`, `y` coordinates and color values.}
}
\value{
Heatmap image.
}
\description{
Generates heatmap image.
}
