% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geerate_explanations.R
\docType{data}
\name{sauron_available_methods}
\alias{sauron_available_methods}
\title{Acronyms and names for available explanation methods.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\usage{
sauron_available_methods
}
\description{
Acronyms and names for available explanation methods.
}
\keyword{datasets}
