% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnn_explanations.R
\name{smooth_grad_x_input}
\alias{smooth_grad_x_input}
\title{Calculates smooth-gradXInput for a CNN.}
\usage{
smooth_grad_x_input(smooth_gradients, input_imgs, preprocessing_function)
}
\arguments{
\item{smooth_gradients}{Smooth gradients.}

\item{input_imgs}{Input images if for of 4-D tensor.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}
}
\value{
smooth-gradXinput for a CNN.
}
\description{
Calculates gradientXinput for a CNN.
}
