% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnn_explanations.R
\name{vanilla_gradient}
\alias{vanilla_gradient}
\title{Calculates vanilla gradient for a CNN.}
\usage{
vanilla_gradient(
  model,
  input_imgs,
  preprocessing_function = NULL,
  class_index = NULL
)
}
\arguments{
\item{model}{Tensorflow model.}

\item{input_imgs}{Input images if for of 4-D tensor.}

\item{preprocessing_function}{Preprocessing function. Default to `NULL`.}

\item{class_index}{Class index. If set to `NULL` index with max predicted probability will be selected.}
}
\description{
Calculates vanilla gradient for a CNN.
}
